Items that are removed are now added the RemovedItems table

+++ oneList/storedProcedures.py
@@ -1,7 +1,8 @@
 from sqlalchemy import text
 #from sqlalchemy.sql.expression import func
 from oneList import db
-from oneList.models import User, Items
+from oneList.models import User, Items, RemovedItems
+from oneList.tools import getEpoch
 
 def selectAllUsers():
     value = text("SELECT * FROM Users") # sqlalchemy
@@ -10,9 +11,18 @@ def selectAllUsers():
 def pairItemAndUser():
     return db.session.query(User, Items).filter(User.uid == Items.addedByUid)
 
-# TODO ADD removed Items to the Removed table
-def removeItem(iid):
-    myitem = db.session.query(Items).filter(Items.iid == iid).first()
-    db.session.delete(myitem)
-    db.session.commit()
+# Moves
+def removeItem(itemID,removedByUid):
+    myitem = db.session.query(Items).filter(Items.iid == itemID).first()
+    if myitem:
+        removedItem = RemovedItems(
+            removedByUid=removedByUid ,
+            addedByUid=myitem.addedByUid ,
+            item=myitem.item ,
+            dateAdded=myitem.dateAdded ,
+            dateRemoved=getEpoch()
+        )
+        db.session.add(removedItem)
+        db.session.delete(myitem)
+        db.session.commit()
